fix : [BE] 관리자 controller 권한 예외처리 추가

+++ backend/src/main/java/finale/bloombloom/api/controller/AdminController.java
@@ -5,10 +5,13 @@ import finale.bloombloom.api.request.AdminUpdateRequest;
 import finale.bloombloom.api.response.StoreDetailResponse;
 import finale.bloombloom.api.response.StoreListResponse;
 import finale.bloombloom.api.service.AdminService;
+import finale.bloombloom.common.auth.BloomUserDetails;
 import finale.bloombloom.common.model.response.Result;
+import finale.bloombloom.db.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -28,7 +31,27 @@ public class AdminController {
      * 작성자 : 박건우
      */
     @GetMapping
-    ResponseEntity<Result> findAllStore() {
+    ResponseEntity<Result> findAllStore(
+            Authentication authentication
+    ) {
+        if (authentication == null)
+            return ResponseEntity.status(401).body(
+                    Result.builder()
+                            .message("인증 실패")
+                            .build()
+            );
+
+        String role = null;
+        for (GrantedAuthority authority : authentication.getAuthorities())
+            role = authority.getAuthority();
+
+        if (!role.equals("ROLE_ADMIN"))
+            return ResponseEntity.status(403).body(
+                    Result.builder()
+                            .message("권한 없음")
+                            .build()
+            );
+
         List<StoreListResponse> stores = adminService.findAllStore();
         return ResponseEntity.status(200).body(
                 Result.builder()
@@ -47,6 +70,21 @@ public class AdminController {
             Authentication authentication,
             @PathVariable Long storeSeq
     ) {
+        if (authentication == null)
+            return ResponseEntity.status(401).body(
+                    Result.builder()
+                            .message("인증 실패")
+                            .build()
+            );
+
+        String role = ((GrantedAuthority) authentication.getAuthorities()).getAuthority();
+        if (!role.equals("ROLE_ADMIN"))
+            return ResponseEntity.status(403).body(
+                    Result.builder()
+                            .message("권한 없음")
+                            .build()
+            );
+
         StoreDetailResponse store = adminService.findStore(storeSeq);
         return ResponseEntity.status(200).body(
                 Result.builder()
@@ -65,6 +103,21 @@ public class AdminController {
             Authentication authentication,
             @RequestParam String storeName
     ) {
+        if (authentication == null)
+            return ResponseEntity.status(401).body(
+                    Result.builder()
+                            .message("인증 실패")
+                            .build()
+            );
+
+        String role = ((GrantedAuthority) authentication.getAuthorities()).getAuthority();
+        if (!role.equals("ROLE_ADMIN"))
+            return ResponseEntity.status(403).body(
+                    Result.builder()
+                            .message("권한 없음")
+                            .build()
+            );
+
         List<StoreListResponse> stores = adminService.searchStore(storeName);
         return ResponseEntity.status(200).body(
                 Result.builder()
@@ -83,6 +136,21 @@ public class AdminController {
             Authentication authentication,
             @PathVariable Long storeSeq
     ) {
+        if (authentication == null)
+            return ResponseEntity.status(401).body(
+                    Result.builder()
+                            .message("인증 실패")
+                            .build()
+            );
+
+        String role = ((GrantedAuthority) authentication.getAuthorities()).getAuthority();
+        if (!role.equals("ROLE_ADMIN"))
+            return ResponseEntity.status(403).body(
+                    Result.builder()
+                            .message("권한 없음")
+                            .build()
+            );
+
         adminService.deleteStore(storeSeq);
         return ResponseEntity.status(200).body(
                 Result.builder()
@@ -101,6 +169,21 @@ public class AdminController {
             @RequestPart(value = "request") AdminSaveRequest req,
             @RequestPart(value = "file") MultipartFile file
     ) {
+        if (authentication == null)
+            return ResponseEntity.status(401).body(
+                    Result.builder()
+                            .message("인증 실패")
+                            .build()
+            );
+
+        String role = ((GrantedAuthority) authentication.getAuthorities()).getAuthority();
+        if (!role.equals("ROLE_ADMIN"))
+            return ResponseEntity.status(403).body(
+                    Result.builder()
+                            .message("권한 없음")
+                            .build()
+            );
+
         String storeImageLink = adminService.saveStore(req, file);
         Map<String, String> response = new HashMap<>();
         response.put("storeImageLink", storeImageLink);
@@ -122,6 +205,21 @@ public class AdminController {
             Authentication authentication,
             @RequestBody AdminUpdateRequest req
     ) {
+        if (authentication == null)
+            return ResponseEntity.status(401).body(
+                    Result.builder()
+                            .message("인증 실패")
+                            .build()
+            );
+
+        String role = ((GrantedAuthority) authentication.getAuthorities()).getAuthority();
+        if (!role.equals("ROLE_ADMIN"))
+            return ResponseEntity.status(403).body(
+                    Result.builder()
+                            .message("권한 없음")
+                            .build()
+            );
+
         adminService.updateStore(req);
         return ResponseEntity.status(200).body(
                 Result.builder()
